+++ scanapi/tree/endpoint_node.py
@@ -1,5 +1,7 @@
+import copy
 import logging
 from itertools import chain
+from typing import Any, Dict, Optional
 
 from scanapi.evaluators import SpecEvaluator
 from scanapi.exit_code import ExitCode
@@ -137,6 +139,38 @@ class EndpointNode:
             [bool]: true if the node has no parent, false otherwise.
         """
         return not self.parent
+
+    def propagate_spec_vars(
+        self,
+        spec_vars: Dict[str, Any],
+        extras: Optional[Dict[str, Any]] = None,
+    ) -> None:
+        """Update the endpoint node spec_vars and propagate those changes
+        to the parent nodes. It only includes new variables.
+
+        Args:
+            new_spec_vars [dic]: the new spec_vars.
+            extras [dict]: extra variables used to update the spec_vars.
+        """
+        new_spec_vars = {
+            key: value
+            for key, value in spec_vars.items()
+            if key not in self.spec_vars.registry
+        }
+        self.spec_vars.update(new_spec_vars, extras)
+        if not self.is_root:
+            self.parent.propagate_spec_vars(spec_vars, extras)
+
+    def get_all_vars(self) -> Dict[str, Any]:
+        """Get all variables in spec_vars from the node and its parents.
+
+        Returns:
+            [dict]: dict from the variable's name to its value.
+        """
+        variables = copy.deepcopy(self.spec_vars.registry)
+        if not self.is_root:
+            variables.update(self.parent.get_all_vars())
+        return variables
 
     def run(self):
         """Run the requests of the node and all children nodes.